Fixed unit test
git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@20994 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ dso-l2/src/test/java/com/tc/objectserver/persistence/api/StringIndexTest.java
@@ -28,27 +28,27 @@ public class StringIndexTest extends TestCase {
     assertNotNull(persistor.loadCalls.poll(0));
 
     // make sure index 0 returns null;
-    assertNull(index.getStringFor(0));
+    assertNull(index.getStringFor(0L));
     
     // make sure it assigns the index as expected
     String testString = ":SLDKFJSD";
-    assertEquals(1, index.getOrCreateIndexFor(testString));
+    assertEquals(1L, index.getOrCreateIndexFor(testString));
     
     // make sure it save the mapping as expected
-    assertEquals(testString, persistor.target.get(1));
+    assertEquals(testString, persistor.target.get(1L));
     
     // clear the map
     persistor.target.clear();
     
     // load up the persistor
     int max = 100;
-    for (int i=1; i<max; i++) {
-      persistor.target.put((long) i, "" + i);
+    for (long i=1; i<max; i++) {
+      persistor.target.put( i, "" + i);
     }
     index = new StringIndexImpl(persistor);
     
     // make sure the mappings are loaded as expected
-    for (int i=1; i<max; i++) {
+    for (long i=1; i<max; i++) {
       String string = "" + i;
       assertEquals(string, index.getStringFor(i));
       assertEquals(i, index.getOrCreateIndexFor(string));
@@ -58,7 +58,7 @@ public class StringIndexTest extends TestCase {
     testString = "lksdfkljdfkjl";
     assertEquals(max, index.getOrCreateIndexFor(testString));
     
-    assertNull(index.getStringFor(0));
+    assertNull(index.getStringFor(0L));
   }
   
   private static final class TestStringIndexPersistor implements StringIndexPersistor {
